Release 0.99.1

+++ pytsite/widget/_select.py
@@ -310,7 +310,7 @@ class Pager(_base.Abstract):
             raise ValueError("'total_items' is required argument.")
 
         self._items_per_page = int(kwargs.get('per_page', 100))
-        self._total_pages = _ceil(self._total_items / self._items_per_page)
+        self._total_pages = int(_ceil(self._total_items / self._items_per_page))
         self._visible_numbers = int(kwargs.get('visible_numbers', 5)) - 1
         self._ajax = kwargs.get('ajax', '')
